make_IndexedStorage
