+++ PySDM/core.py
@@ -8,6 +8,7 @@ from PySDM.state.particles import Particles
 from PySDM.storages.index import make_Index
 from PySDM.storages.pair_indicator import make_PairIndicator
 from PySDM.storages.pairwise_storage import make_PairwiseStorage
+from PySDM.storages.indexed_storage import make_IndexedStorage
 
 
 class Core:
@@ -30,6 +31,7 @@ class Core:
         self.Index = make_Index(backend)
         self.PairIndicator = make_PairIndicator(backend)
         self.PairwiseStorage = make_PairwiseStorage(backend)
+        self.IndexedStorage = make_IndexedStorage(backend)
 
     @property
     def env(self):
@@ -42,10 +44,6 @@ class Core:
     @property
     def Storage(self):
         return self.backend.Storage
-
-    @property
-    def IndexedStorage(self):
-        return self.backend.IndexedStorage
 
     @property
     def Random(self):
